Migrate GNETextSearch to Swift Package Manager
- Add Package.swift with C library and test targets
- Restructure sources, public headers, and test resources for SwiftPM
- Replace Xcode CI with swift test and document SwiftPM usage
- Add Swift import coverage for the C module
