--- GNETextSearch.xcodeproj/xcshareddata/xcbaselines/5711A7F21B949E440088910A.xcbaseline/Info.plist
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>runDestinationsByUUID</key>
-	<dict>
-		<key>57510002-C85F-4DFF-A2CB-FB4008726F42</key>
-		<dict>
-			<key>localComputer</key>
-			<dict>
-				<key>busSpeedInMHz</key>
-				<integer>100</integer>
-				<key>cpuCount</key>
-				<integer>1</integer>
-				<key>cpuKind</key>
-				<string>Intel Core i7</string>
-				<key>cpuSpeedInMHz</key>
-				<integer>2500</integer>
-				<key>logicalCPUCoresPerPackage</key>
-				<integer>8</integer>
-				<key>modelCode</key>
-				<string>MacBookPro11,2</string>
-				<key>physicalCPUCoresPerPackage</key>
-				<integer>4</integer>
-				<key>platformIdentifier</key>
-				<string>com.apple.platform.macosx</string>
-			</dict>
-			<key>targetArchitecture</key>
-			<string>x86_64</string>
-		</dict>
-	</dict>
-</dict>
-</plist>
