+++ Sources/GNETextSearch/include/GNETextSearch/GNETextSearch.h
@@ -1,0 +1,8 @@
+#ifndef GNETextSearch_h
+#define GNETextSearch_h
+
+#include <GNETextSearch/Types.h>
+#include <GNETextSearch/CountedSet.h>
+#include <GNETextSearch/TernaryTree.h>
+
+#endif /* GNETextSearch_h */
