+++ Sources/GNETextSearch/include/GNETextSearch/Types.h
@@ -1,0 +1,24 @@
+//
+//  Types.h
+//  GNETextSearch
+//
+//  Created by Anthony Drendel on 1/9/16.
+//  Copyright © 2016 Gone East LLC. All rights reserved.
+//
+
+#ifndef GNETextSearchTypes_h
+#define GNETextSearchTypes_h
+
+#include <stdlib.h>
+#include <stdint.h>
+#include <stdbool.h>
+
+#define failure 0
+#define success 1
+
+#define result _Result
+typedef int _Result;
+
+typedef int64_t GNEInteger;
+
+#endif /* GNETextSearchTypes_h */
