+++ Tests/GNETextSearchCTests/GNETextSearchTestResources.h
@@ -1,0 +1,3 @@
+#import <Foundation/Foundation.h>
+
+NSBundle *GNETextSearchTestResourcesBundle(void);
