+++ Tests/GNETextSearchCTests/Resources/words-beginning-with-A.txt
